Rename `writerCommand.ts` to `payload.ts`
