+++ client/ts/src/framer/codec.spec.ts
@@ -21,8 +21,8 @@ import {
   WSWriterCodec,
 } from "@/framer/codec";
 import { Frame } from "@/framer/frame";
+import { WriterCommand } from "@/framer/payload";
 import { type WriteRequest } from "@/framer/writer";
-import { WriterCommand } from "@/framer/writerCommand";
 
 describe("encoder", () => {
   describe("base codec", () => {
