+++ client/ts/src/framer/codec.ts
@@ -20,9 +20,9 @@ import { type z } from "zod";
 import { type channel } from "@/channel";
 import { ValidationError } from "@/errors";
 import { type Frame, type Payload } from "@/framer/frame";
+import { WriterCommand } from "@/framer/payload";
 import { type StreamerResponse } from "@/framer/streamer";
 import { type WriteRequest } from "@/framer/writer";
-import { WriterCommand } from "@/framer/writerCommand";
 
 const seriesPldLength = (series: SeriesPayload): number =>
   series.data.byteLength / series.dataType.density.valueOf();
